Random options cannot be used together
Added mutual exclusivity to the random password generation commands. Random commands do not interfere with keyword commands, so mutual exclusivity is not necessary there. Not the safest code but it works. Still need to update README and attempt to determine SystemExit: 2 error with using multiple random commands under mutual exclusion.